Use CommonJS Next.js config for Node 14
