+++ next.config.js
@@ -8,4 +8,4 @@ const nextConfig = {
   },
 };
 
-export default nextConfig;
+module.exports = nextConfig;
